Split into files, extra test for KvStore
