+++ crates/state/src/lib.rs
@@ -1,4 +1,5 @@
-pub mod api;
+pub mod cascade;
+pub mod env;
 pub mod error;
-pub mod lmdb;
+pub mod store;
 pub mod workspace;
