--- crates/state/src/lmdb/mod.rs
@@ -1,3 +1,0 @@
-use lmdb::Environment;
-
-
